+++ packages/discovery-provider/plugins/pedalboard/apps/trending-challenge-rewards/src/sdk.ts
@@ -1,0 +1,8 @@
+import { sdk } from '@audius/sdk' 
+
+export const audiusSdk = () => {
+  return sdk({
+    appName: 'trending-challenge-rewards',
+    environment: process.env.ENVIRONMENT as 'development' | 'staging' | 'production' | undefined ?? 'staging',
+  })
+}
