+++ packages/discovery-provider/plugins/pedalboard/apps/trending-challenge-rewards/src/slack.ts
@@ -6,7 +6,7 @@ import { AlignmentEnum, AsciiTable3 } from 'ascii-table3'
 import { App } from '@pedalboard/basekit'
 import { Err, Ok, Result } from 'ts-results'
 import { SharedData } from './config'
-import { onDisburse } from './app'
+import { onDisburse } from './rewards'
 import { ChallengeDisbursementUserbankFriendly } from './queries'
 import { announceTopFiveTrending } from './trending'
 import { isValidDate } from './utils'
